perf: use a composable rather than BTransition component, eliminating a useless fragment
