+++ .github/ISSUE_TEMPLATE/compatibility_report.yaml
@@ -32,21 +32,3 @@ body:
       placeholder: Feature description
     validations:
       required: true
-  - type: checkboxes
-    id: checkboxes
-    attributes:
-      label: Validations
-      description: Before submitting the issue, please make sure you do the following
-      options:
-        - label: Have tested with the latest version. This is still alpha version and sometime things change rapidly.
-          required: true
-        - label: Follow our [Code of Conduct](https://github.com/bootstrap-vue-next/bootstrap-vue-next/blob/main/CODE_OF_CONDUCT.md)
-          required: true
-        - label: Read the [Contributing Guide](https://github.com/bootstrap-vue-next/bootstrap-vue-next/blob/main/CONTRIBUTING.md).
-          required: true
-        - label: Check that there isn't already an issue that reports the same bug to avoid creating a duplicate.
-          required: true
-        - label: Check that this is a concrete bug. For Q&A, please open a GitHub Discussion instead.
-          required: true
-        - label: The provided link is to the documentation of the missing feature.
-          required: true
