+++ .github/ISSUE_TEMPLATE/feature_request.yaml
@@ -30,15 +30,3 @@ body:
     attributes:
       label: Additional context
       description: Any other context or screenshots about the feature request here.
-  - type: checkboxes
-    id: checkboxes
-    attributes:
-      label: Validations
-      description: Before submitting the issue, please make sure you do the following
-      options:
-        - label: Follow our [Code of Conduct](https://github.com/bootstrap-vue-next/bootstrap-vue-next/blob/main/CODE_OF_CONDUCT.md)
-          required: true
-        - label: Read the [Contributing Guide](https://github.com/bootstrap-vue-next/bootstrap-vue-next/blob/main/CONTRIBUTING.md).
-          required: true
-        - label: Check that there isn't already an issue that request the same feature to avoid creating a duplicate.
-          required: true
