+++ packages/bootstrap-vue-next/src/styles/styles.scss
@@ -51,3 +51,13 @@
     }
   }
 }
+
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.25s linear;
+}
+
+.fade-enter-from,
+.fade-leave-to {
+  opacity: 0;
+}
